requirements.txt
sqlparse==0.5.0
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
black==24.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.5
defusedxml==0.7.1

gunicorn==20.1.0
daphne==4.0.0
dj-rest-auth==4.0.0
Django==5.0.5
django-environ==0.7.0
django-storages==1.12.2
django_crontab==0.7.1
django-allauth==0.52.0
django-cors-headers==4.3.1

django-cron==0.6.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1


hyperlink==21.0.0
idna==3.6
incremental==22.10.0
kombu==5.3.7
mypy-extensions==1.0.0
oauthlib==3.2.2



packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
pycryptodome

requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
